restructure package.json script
